+++ packages/pyapprox/src/pyapprox/pde/galerkin/postprocessing/__init__.py
@@ -6,13 +6,15 @@ Organized by physics type:
 """
 
 from pyapprox.pde.galerkin.postprocessing.elasticity import (
-    strain_from_displacement_2d,
-    stress_from_strain_2d,
-    von_mises_stress_2d,
+    integrate,
+    strain_from_displacement,
+    stress_from_strain,
+    von_mises_stress,
 )
 
 __all__ = [
-    "von_mises_stress_2d",
-    "strain_from_displacement_2d",
-    "stress_from_strain_2d",
+    "integrate",
+    "strain_from_displacement",
+    "stress_from_strain",
+    "von_mises_stress",
 ]